Make pydantic usage more realistic.
The DimensionGroup toy now works, and the workspace config stuff
probably does (since it's based on ExtensionConfig, which I did test).
Much of the rest doesn't work, but at least I think that's "just" due
to me not understanding pydantic_core fully, rather than because it's
fundamentally impossible for pydantic to do (as was the case before).